+++ crates/tabby-common/src/api/event.rs
@@ -1,4 +1,3 @@
-use async_trait::async_trait;
 use serde::{Deserialize, Serialize};
 use utoipa::ToSchema;
 
@@ -93,9 +92,8 @@ pub struct Segments {
     pub clipboard: Option<String>,
 }
 
-#[async_trait]
 pub trait EventLogger: Send + Sync {
-    async fn log(&self, e: Event);
+    fn log(&self, e: Event);
 }
 
 #[derive(Serialize)]
@@ -108,9 +106,8 @@ pub trait RawEventLogger: Send + Sync {
     fn log(&self, content: String);
 }
 
-#[async_trait]
 impl<T: RawEventLogger> EventLogger for T {
-    async fn log(&self, e: Event) {
+    fn log(&self, e: Event) {
         let content = serdeconv::to_json_string(&Log {
             ts: timestamp(),
             event: e,
