+++ ee/tabby-webserver/src/service/email/mod.rs
@@ -17,6 +17,7 @@ use tracing::warn;
 mod templates;
 #[cfg(test)]
 pub mod testutils;
+mod user_completions;
 
 use crate::schema::{
     email::{
